attendance-system/backend/requirements.txt
fastapi
uvicorn
python-multipart
boto3
face_recognition
numpy
pillow
python-dotenv
pydantic
